Seperate IB verbs functionality in a file
